Cambiar nota a double

+++ src/main/java/fb2/studentanalytics/model/Student.java
@@ -12,9 +12,9 @@ public class Student {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private int averageGrade;
+    private double averageGrade;
 
-    public Student(String name, int averageGrade) {
+    public Student(String name, double averageGrade) {
         this.name = name;
         this.averageGrade = averageGrade;
     }
